UGLY HACK - REMOVE ASAP: Restore login via auth-proxy
The 5.x behavior of Grafana was to implicitly log in the user
when using auth-proxy for a single request.

In 6.x this is no longer the case. The head for auth-proxy has to
be passed on each request.

Restore the old behavior until 'Automatic login by token url #3752'
is ready or we implement our own jwt based solution.